refactor(frontend): split apollo route composition
