+++ frontend/apps/apollo/src/features/assignments/mock-data.tsx
@@ -1,0 +1,14 @@
+import { FileCode2, SquareTerminal } from "lucide-react";
+
+export const starterFiles = [
+  {
+    name: "linked_list.py",
+    meta: "Python Source • 2.4 KB",
+    icon: <FileCode2 size={18} />,
+  },
+  {
+    name: "test_suite.py",
+    meta: "Python Test • 5.1 KB",
+    icon: <SquareTerminal size={18} />,
+  },
+] as const;
