+++ frontend/apps/apollo/src/features/shell/ProtectedAppLayout.tsx
@@ -1,0 +1,13 @@
+import { Outlet } from "@tanstack/react-router";
+import { AppFrame } from "@optimark/calliope";
+
+import { AppSidebar } from "./AppSidebar";
+import { AppTopbar } from "./AppTopbar";
+
+export function ProtectedAppLayout() {
+  return (
+    <AppFrame sidebar={<AppSidebar />} topbar={<AppTopbar />}>
+      <Outlet />
+    </AppFrame>
+  );
+}
