+++ frontend/apps/apollo/src/main.tsx
@@ -1,1296 +1,11 @@
-import React, { useMemo, useState } from "react";
+import React from "react";
 import ReactDOM from "react-dom/client";
-import {
-  QueryClient,
-  QueryClientProvider,
-  useMutation,
-  useQuery,
-} from "@tanstack/react-query";
-import {
-  Link,
-  Outlet,
-  RouterProvider,
-  createRootRouteWithContext,
-  createRoute,
-  createRouter,
-  redirect,
-  useNavigate,
-  useRouterState,
-} from "@tanstack/react-router";
-import {
-  Bell,
-  BookOpen,
-  ChartColumn,
-  ChevronDown,
-  CircleAlert,
-  Code2,
-  Download,
-  FileCode2,
-  Filter,
-  FolderOpen,
-  Gauge,
-  GraduationCap,
-  History,
-  LayoutDashboard,
-  ListChecks,
-  LoaderCircle,
-  LogOut,
-  Search,
-  Settings,
-  Sparkles,
-  SquareTerminal,
-  Upload,
-  Users,
-} from "lucide-react";
-import {
-  AppFrame,
-  BottomActionBar,
-  BrandLockup,
-  EmptyState,
-  FormFieldScaffold,
-  MetricCard,
-  PageHeader,
-  PageShell,
-  SectionHeading,
-  SidebarNavItem,
-  SidebarShell,
-  StatusPill,
-  SurfacePanel,
-  Topbar,
-  TopbarTab,
-  brand,
-  sidebarUtilityLinks,
-  topTabs,
-} from "@optimark/calliope";
+import { QueryClientProvider } from "@tanstack/react-query";
+import { RouterProvider } from "@tanstack/react-router";
+
 import "@optimark/calliope/system.css";
 import "./styles.css";
-
-type AppContext = {
-  queryClient: QueryClient;
-};
-
-type AuthSearch = {
-  redirect?: string;
-};
-
-type SessionUser = {
-  id: string;
-  email: string;
-  display_name: string;
-};
-
-type SessionResponse = {
-  user: SessionUser;
-  provider: string;
-  expires_at: string;
-};
-
-type AuthPayload = {
-  email: string;
-  password: string;
-};
-
-type SignupPayload = AuthPayload & {
-  display_name: string;
-};
-
-class ApiError extends Error {
-  status: number;
-
-  constructor(status: number, message: string) {
-    super(message);
-    this.name = "ApiError";
-    this.status = status;
-  }
-}
-
-const queryClient = new QueryClient();
-const sessionQueryKey = ["auth", "session"] as const;
-
-const navItems = [
-  { to: "/dashboard", label: "Dashboard", icon: LayoutDashboard },
-  { to: "/assignments", label: "Assignments", icon: BookOpen },
-  { to: "/submissions", label: "Submissions", icon: ListChecks },
-  { to: "/gradebook", label: "Gradebook", icon: ChartColumn },
-  { to: "/students", label: "Students", icon: Users },
-  { to: "/settings", label: "Settings", icon: Settings },
-] as const;
-
-const assignmentRows = [
-  {
-    name: "Homework 4: Linked Lists",
-    type: "Practical",
-    status: { label: "Published", tone: "primary" as const },
-    due: "Oct 12, 23:59",
-    submissions: "42 / 45",
-  },
-  {
-    name: "Assignment 3: Binary Trees",
-    type: "Programming",
-    status: { label: "Reviewing", tone: "secondary" as const },
-    due: "Oct 05, 12:00",
-    submissions: "45 / 45",
-  },
-  {
-    name: "Midterm Quiz: Core Concepts",
-    type: "Exam",
-    status: { label: "Draft", tone: "default" as const },
-    due: "Oct 24, 09:00",
-    submissions: "0 / 45",
-  },
-  {
-    name: "Homework 5: Graph Theory",
-    type: "Practical",
-    status: { label: "Draft", tone: "default" as const },
-    due: "Nov 02, 23:59",
-    submissions: "0 / 45",
-  },
-] as const;
-
-const activityFeed = [
-  {
-    title: "Student A submitted Homework 4",
-    when: "12 minutes ago",
-    detail: "New programming artifact is ready for queueing.",
-    tone: "primary" as const,
-  },
-  {
-    title: "Autograde completed for Assignment 3",
-    when: "45 minutes ago",
-    detail: "82% average • 4 error flags surfaced for review.",
-    tone: "secondary" as const,
-  },
-  {
-    title: "Manual review started on Midterm Essays",
-    when: "2 hours ago",
-    detail: "TA review batch is currently in progress.",
-    tone: "default" as const,
-  },
-  {
-    title: "System: Plagiarism detected in Homework 4",
-    when: "3 hours ago",
-    detail: "One critical alert needs instructor verification.",
-    tone: "danger" as const,
-  },
-] as const;
-
-const starterFiles = [
-  {
-    name: "linked_list.py",
-    meta: "Python Source • 2.4 KB",
-    icon: <FileCode2 size={18} />,
-  },
-  {
-    name: "test_suite.py",
-    meta: "Python Test • 5.1 KB",
-    icon: <SquareTerminal size={18} />,
-  },
-] as const;
-
-const gradebookRows = [
-  {
-    initials: "AA",
-    student: "Alex Abramov",
-    email: "a.abramov@university.edu",
-    accent: "secondary" as const,
-    scores: [
-      { value: "95/100", status: "Released", tone: "primary" as const },
-      { value: "88/100", status: "Released", tone: "primary" as const },
-      { value: "--/100", status: "Pending", tone: "secondary" as const },
-      { value: "--/100", status: "Missing", tone: "default" as const },
-      { value: "142/150", status: "Released", tone: "primary" as const },
-    ],
-    overall: "91.4%",
-    overallTone: "primary" as const,
-  },
-  {
-    initials: "EL",
-    student: "Elena Laveau",
-    email: "e.laveau@university.edu",
-    accent: "secondary" as const,
-    scores: [
-      { value: "100/100", status: "Released", tone: "primary" as const },
-      { value: "94/100", status: "Released", tone: "primary" as const },
-      { value: "98/100", status: "Released", tone: "primary" as const },
-      { value: "92/100", status: "Released", tone: "primary" as const },
-      { value: "148/150", status: "Released", tone: "primary" as const },
-    ],
-    overall: "97.8%",
-    overallTone: "primary" as const,
-  },
-  {
-    initials: "JM",
-    student: "Jordan Miller",
-    email: "j.miller@university.edu",
-    accent: "danger" as const,
-    scores: [
-      { value: "64/100", status: "Released", tone: "primary" as const },
-      { value: "72/100", status: "Released", tone: "primary" as const },
-      { value: "--/100", status: "Pending", tone: "secondary" as const },
-      { value: "--/100", status: "Pending", tone: "secondary" as const },
-      { value: "110/150", status: "Released", tone: "primary" as const },
-    ],
-    overall: "68.2%",
-    overallTone: "danger" as const,
-  },
-  {
-    initials: "SW",
-    student: "Sarah Wu",
-    email: "s.wu@university.edu",
-    accent: "secondary" as const,
-    scores: [
-      { value: "92/100", status: "Released", tone: "primary" as const },
-      { value: "89/100", status: "Released", tone: "primary" as const },
-      { value: "91/100", status: "Released", tone: "primary" as const },
-      { value: "88/100", status: "Released", tone: "primary" as const },
-      { value: "138/150", status: "Released", tone: "primary" as const },
-    ],
-    overall: "90.5%",
-    overallTone: "primary" as const,
-  },
-] as const;
-
-const gradeDistribution = [
-  { label: "F", height: "18%", accent: false },
-  { label: "D", height: "42%", accent: false },
-  { label: "C", height: "78%", accent: true },
-  { label: "B", height: "60%", accent: false },
-  { label: "A", height: "14%", accent: false },
-] as const;
-
-async function requestJson<T>(path: string, init?: RequestInit): Promise<T> {
-  const response = await fetch(path, {
-    ...init,
-    credentials: "include",
-    headers: {
-      "Content-Type": "application/json",
-      ...(init?.headers ?? {}),
-    },
-  });
-
-  if (!response.ok) {
-    let detail = "Request failed.";
-
-    try {
-      const body = (await response.json()) as { detail?: string };
-      detail = body.detail ?? detail;
-    } catch {
-      detail = response.statusText || detail;
-    }
-
-    throw new ApiError(response.status, detail);
-  }
-
-  return (await response.json()) as T;
-}
-
-async function fetchSession(): Promise<SessionResponse | null> {
-  try {
-    return await requestJson<SessionResponse>("/api/v1/auth/session", {
-      method: "GET",
-    });
-  } catch (error) {
-    if (error instanceof ApiError && error.status === 401) {
-      return null;
-    }
-
-    throw error;
-  }
-}
-
-async function loginRequest(payload: AuthPayload): Promise<SessionResponse> {
-  return requestJson<SessionResponse>("/api/v1/auth/login", {
-    method: "POST",
-    body: JSON.stringify(payload),
-  });
-}
-
-async function signupRequest(payload: SignupPayload): Promise<SessionResponse> {
-  return requestJson<SessionResponse>("/api/v1/auth/signup", {
-    method: "POST",
-    body: JSON.stringify(payload),
-  });
-}
-
-async function logoutRequest(): Promise<void> {
-  const response = await fetch("/api/v1/auth/logout", {
-    method: "POST",
-    credentials: "include",
-  });
-
-  if (!response.ok) {
-    throw new ApiError(response.status, "Unable to end the current session.");
-  }
-}
-
-function sessionQueryOptions() {
-  return {
-    queryKey: sessionQueryKey,
-    queryFn: fetchSession,
-    staleTime: 60_000,
-    retry: false,
-  };
-}
-
-async function ensureSession(queryClient: QueryClient): Promise<SessionResponse | null> {
-  return queryClient.ensureQueryData(sessionQueryOptions());
-}
-
-async function requireAuthenticated(queryClient: QueryClient, redirectPath: string) {
-  const session = await ensureSession(queryClient);
-
-  if (!session) {
-    throw redirect({
-      to: "/login",
-      search: { redirect: redirectPath },
-    });
-  }
-
-  return session;
-}
-
-async function redirectAuthenticated(queryClient: QueryClient) {
-  const session = await ensureSession(queryClient);
-
-  if (session) {
-    throw redirect({ to: "/dashboard" });
-  }
-}
-
-const rootRoute = createRootRouteWithContext<AppContext>()({
-  component: RootLayout,
-});
-
-const indexRoute = createRoute({
-  getParentRoute: () => rootRoute,
-  path: "/",
-  beforeLoad: async ({ context }) => {
-    const session = await ensureSession(context.queryClient);
-    throw redirect({ to: session ? "/dashboard" : "/login" });
-  },
-});
-
-const loginRoute = createRoute({
-  getParentRoute: () => rootRoute,
-  path: "/login",
-  validateSearch: (search: Record<string, unknown>): AuthSearch => ({
-    redirect: typeof search.redirect === "string" ? search.redirect : undefined,
-  }),
-  beforeLoad: async ({ context }) => {
-    await redirectAuthenticated(context.queryClient);
-  },
-  component: LoginPage,
-});
-
-const signupRoute = createRoute({
-  getParentRoute: () => rootRoute,
-  path: "/signup",
-  validateSearch: (search: Record<string, unknown>): AuthSearch => ({
-    redirect: typeof search.redirect === "string" ? search.redirect : undefined,
-  }),
-  beforeLoad: async ({ context }) => {
-    await redirectAuthenticated(context.queryClient);
-  },
-  component: SignupPage,
-});
-
-const dashboardRoute = createRoute({
-  getParentRoute: () => rootRoute,
-  path: "/dashboard",
-  beforeLoad: async ({ context, location }) => {
-    await requireAuthenticated(context.queryClient, location.pathname);
-  },
-  component: DashboardPage,
-});
-
-const assignmentsRoute = createRoute({
-  getParentRoute: () => rootRoute,
-  path: "/assignments",
-  beforeLoad: async ({ context, location }) => {
-    await requireAuthenticated(context.queryClient, location.pathname);
-  },
-  component: AssignmentsPage,
-});
-
-const assignmentEditorRoute = createRoute({
-  getParentRoute: () => rootRoute,
-  path: "/assignments/new",
-  beforeLoad: async ({ context, location }) => {
-    await requireAuthenticated(context.queryClient, location.pathname);
-  },
-  component: AssignmentBuilderPage,
-});
-
-const submissionsRoute = createRoute({
-  getParentRoute: () => rootRoute,
-  path: "/submissions",
-  beforeLoad: async ({ context, location }) => {
-    await requireAuthenticated(context.queryClient, location.pathname);
-  },
-  component: SubmissionsPage,
-});
-
-const gradebookRoute = createRoute({
-  getParentRoute: () => rootRoute,
-  path: "/gradebook",
-  beforeLoad: async ({ context, location }) => {
-    await requireAuthenticated(context.queryClient, location.pathname);
-  },
-  component: GradebookPage,
-});
-
-const studentsRoute = createRoute({
-  getParentRoute: () => rootRoute,
-  path: "/students",
-  beforeLoad: async ({ context, location }) => {
-    await requireAuthenticated(context.queryClient, location.pathname);
-  },
-  component: StudentsPage,
-});
-
-const settingsRoute = createRoute({
-  getParentRoute: () => rootRoute,
-  path: "/settings",
-  beforeLoad: async ({ context, location }) => {
-    await requireAuthenticated(context.queryClient, location.pathname);
-  },
-  component: SettingsPage,
-});
-
-const routeTree = rootRoute.addChildren([
-  indexRoute,
-  loginRoute,
-  signupRoute,
-  dashboardRoute,
-  assignmentsRoute,
-  assignmentEditorRoute,
-  submissionsRoute,
-  gradebookRoute,
-  studentsRoute,
-  settingsRoute,
-]);
-
-const router = createRouter({
-  routeTree,
-  context: {
-    queryClient,
-  },
-  defaultPreload: "intent",
-});
-
-declare module "@tanstack/react-router" {
-  interface Register {
-    router: typeof router;
-  }
-}
-
-function RootLayout() {
-  const pathname = useRouterState({
-    select: (state) => state.location.pathname,
-  });
-  const isAuthRoute = pathname === "/login" || pathname === "/signup";
-
-  if (isAuthRoute) {
-    return <AuthViewport />;
-  }
-
-  return (
-    <AppFrame sidebar={<AppSidebar />} topbar={<AppTopbar />}>
-      <Outlet />
-    </AppFrame>
-  );
-}
-
-function AuthViewport() {
-  return (
-    <div className="app-auth-shell">
-      <div className="app-auth-background" />
-      <div className="app-auth-grid">
-        <section className="app-auth-story">
-          <BrandLockup
-            name={brand.name}
-            context={`${brand.courseLabel} • ${brand.courseTerm}`}
-            mark={<Code2 size={18} />}
-          />
-          <div className="app-auth-copy">
-            <span className="app-smallcaps">Hosted Access</span>
-            <h2>The academic workspace with operational calm.</h2>
-            <p>
-              Sign in to restore your course session, review assessment activity,
-              and continue in the same curated shell used across the app.
-            </p>
-          </div>
-          <div className="app-auth-highlights">
-            <SurfacePanel muted className="app-auth-note">
-              <strong>Session restoration</strong>
-              <p>Cookie-backed app access restores your workspace on reload.</p>
-            </SurfacePanel>
-            <SurfacePanel muted className="app-auth-note">
-              <strong>Protected surfaces</strong>
-              <p>Instructor and student routes stay gated until a valid session exists.</p>
-            </SurfacePanel>
-          </div>
-        </section>
-        <div className="app-auth-panel">
-          <Outlet />
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function AppSidebar() {
-  const { data: session } = useQuery(sessionQueryOptions());
-
-  return (
-    <SidebarShell
-      brand={
-        <BrandLockup
-          name={brand.name}
-          context={brand.courseLabel}
-          mark={<Code2 size={18} />}
-        />
-      }
-      navigation={navItems.map(({ to, label, icon: Icon }) => (
-        <Link key={to} to={to}>
-          {({ isActive }) => (
-            <SidebarNavItem
-              active={isActive}
-              icon={<Icon size={20} />}
-              label={label}
-            />
-          )}
-        </Link>
-      ))}
-      primaryAction={
-        <Link to="/assignments/new" className="app-primary-action">
-          New Assessment
-        </Link>
-      }
-      utilityLinks={sidebarUtilityLinks.map((item) => (
-        <div key={item.key} className="app-utility-link">
-          {item.label}
-        </div>
-      ))}
-      profile={
-        <SidebarProfile
-          displayName={session?.user.display_name ?? brand.instructorName}
-          email={session?.user.email ?? "staff@optimark.dev"}
-        />
-      }
-    />
-  );
-}
-
-function SidebarProfile({
-  displayName,
-  email,
-}: {
-  displayName: string;
-  email: string;
-}) {
-  const navigate = useNavigate();
-  const logout = useMutation({
-    mutationFn: logoutRequest,
-    onSuccess: async () => {
-      queryClient.setQueryData(sessionQueryKey, null);
-      await queryClient.invalidateQueries({ queryKey: sessionQueryKey });
-      await navigate({ to: "/login" });
-    },
-  });
-  const initials = useMemo(
-    () =>
-      displayName
-        .split(" ")
-        .map((part) => part[0])
-        .join("")
-        .slice(0, 2)
-        .toUpperCase(),
-    [displayName],
-  );
-
-  return (
-    <div className="app-profile-stack">
-      <div className="app-profile-chip">
-        <div className="app-profile-avatar">{initials}</div>
-        <div>
-          <strong>{displayName}</strong>
-          <p>{email}</p>
-        </div>
-      </div>
-      <button
-        className="app-secondary-action app-sidebar-logout"
-        type="button"
-        onClick={() => logout.mutate()}
-        disabled={logout.isPending}
-      >
-        {logout.isPending ? <LoaderCircle size={16} className="app-spin" /> : <LogOut size={16} />}
-        Log Out
-      </button>
-      {logout.isError ? (
-        <p className="app-inline-error">We could not end the current session. Try again.</p>
-      ) : null}
-    </div>
-  );
-}
-
-function AppTopbar() {
-  const pathname = useRouterState({
-    select: (state) => state.location.pathname,
-  });
-  const { data: session } = useQuery(sessionQueryOptions());
-  const activeTopTab = pathname === "/gradebook" ? "analytics" : "course-settings";
-  const initials = useMemo(
-    () =>
-      (session?.user.display_name ?? brand.instructorName)
-        .split(" ")
-        .map((part) => part[0])
-        .join("")
-        .slice(0, 2)
-        .toUpperCase(),
-    [session?.user.display_name],
-  );
-
-  return (
-    <Topbar
-      tabs={topTabs.map((tab) => (
-        <TopbarTab key={tab.key} active={tab.key === activeTopTab}>
-          {tab.label}
-        </TopbarTab>
-      ))}
-      search={
-        <label className="app-search">
-          <Search size={16} />
-          <input placeholder="Search assessments..." />
-        </label>
-      }
-      tools={
-        <div className="app-topbar-tools">
-          <button className="app-icon-button" type="button" aria-label="Notifications">
-            <Bell size={18} />
-          </button>
-          <button className="app-icon-button" type="button" aria-label="History">
-            <History size={18} />
-          </button>
-          <div className="app-topbar-session">
-            <span>{session?.user.display_name ?? brand.instructorName}</span>
-            <div className="app-topbar-avatar">{initials}</div>
-          </div>
-        </div>
-      }
-    />
-  );
-}
-
-function LoginPage() {
-  const navigate = useNavigate({ from: "/login" });
-  const search = loginRoute.useSearch();
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const login = useMutation({
-    mutationFn: loginRequest,
-    onSuccess: async (session) => {
-      queryClient.setQueryData(sessionQueryKey, session);
-      await navigate({ to: search.redirect || "/dashboard" });
-    },
-  });
-
-  async function onSubmit(event: React.FormEvent<HTMLFormElement>) {
-    event.preventDefault();
-    login.mutate({ email, password });
-  }
-
-  return (
-    <AuthCard
-      eyebrow="Sign In"
-      title="Resume your hosted workspace."
-      subtitle="Use the account created through the backend auth foundation to restore your session."
-      footer={
-        <p className="app-auth-footer">
-          Need an account?{" "}
-          <Link to="/signup" search={{ redirect: search.redirect }}>
-            Create one
-          </Link>
-        </p>
-      }
-    >
-      <form className="app-auth-form" onSubmit={onSubmit}>
-        <FormFieldScaffold label="Email Address">
-          <input
-            className="app-auth-input"
-            type="email"
-            autoComplete="email"
-            value={email}
-            onChange={(event) => setEmail(event.target.value)}
-            placeholder="instructor@university.edu"
-            required
-          />
-        </FormFieldScaffold>
-        <FormFieldScaffold label="Password">
-          <input
-            className="app-auth-input"
-            type="password"
-            autoComplete="current-password"
-            value={password}
-            onChange={(event) => setPassword(event.target.value)}
-            placeholder="At least 12 characters"
-            required
-          />
-        </FormFieldScaffold>
-        {login.isError ? (
-          <AuthErrorBanner error={login.error} />
-        ) : null}
-        <button className="app-primary-action app-auth-submit" type="submit" disabled={login.isPending}>
-          {login.isPending ? <LoaderCircle size={16} className="app-spin" /> : null}
-          Sign In
-        </button>
-      </form>
-    </AuthCard>
-  );
-}
-
-function SignupPage() {
-  const navigate = useNavigate({ from: "/signup" });
-  const search = signupRoute.useSearch();
-  const [displayName, setDisplayName] = useState("");
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const signup = useMutation({
-    mutationFn: signupRequest,
-    onSuccess: async (session) => {
-      queryClient.setQueryData(sessionQueryKey, session);
-      await navigate({ to: search.redirect || "/dashboard" });
-    },
-  });
-
-  async function onSubmit(event: React.FormEvent<HTMLFormElement>) {
-    event.preventDefault();
-    signup.mutate({ display_name: displayName, email, password });
-  }
-
-  return (
-    <AuthCard
-      eyebrow="Create Account"
-      title="Open a new course session."
-      subtitle="This initial hosted flow stays intentionally simple while preserving the backend seam for future SSO."
-      footer={
-        <p className="app-auth-footer">
-          Already have access?{" "}
-          <Link to="/login" search={{ redirect: search.redirect }}>
-            Sign in
-          </Link>
-        </p>
-      }
-    >
-      <form className="app-auth-form" onSubmit={onSubmit}>
-        <FormFieldScaffold label="Display Name">
-          <input
-            className="app-auth-input"
-            type="text"
-            autoComplete="name"
-            value={displayName}
-            onChange={(event) => setDisplayName(event.target.value)}
-            placeholder="Dr. Aris Thorne"
-            required
-          />
-        </FormFieldScaffold>
-        <FormFieldScaffold label="Email Address">
-          <input
-            className="app-auth-input"
-            type="email"
-            autoComplete="email"
-            value={email}
-            onChange={(event) => setEmail(event.target.value)}
-            placeholder="instructor@university.edu"
-            required
-          />
-        </FormFieldScaffold>
-        <FormFieldScaffold label="Password" support="Minimum 12 characters.">
-          <input
-            className="app-auth-input"
-            type="password"
-            autoComplete="new-password"
-            value={password}
-            onChange={(event) => setPassword(event.target.value)}
-            placeholder="Choose a strong password"
-            required
-          />
-        </FormFieldScaffold>
-        {signup.isError ? (
-          <AuthErrorBanner error={signup.error} />
-        ) : null}
-        <button className="app-primary-action app-auth-submit" type="submit" disabled={signup.isPending}>
-          {signup.isPending ? <LoaderCircle size={16} className="app-spin" /> : null}
-          Create Account
-        </button>
-      </form>
-    </AuthCard>
-  );
-}
-
-function AuthCard({
-  eyebrow,
-  title,
-  subtitle,
-  children,
-  footer,
-}: {
-  eyebrow: string;
-  title: string;
-  subtitle: string;
-  children: React.ReactNode;
-  footer?: React.ReactNode;
-}) {
-  return (
-    <SurfacePanel className="app-auth-card">
-      <div className="app-auth-card-header">
-        <span className="app-smallcaps">{eyebrow}</span>
-        <h1>{title}</h1>
-        <p>{subtitle}</p>
-      </div>
-      {children}
-      {footer}
-    </SurfacePanel>
-  );
-}
-
-function AuthErrorBanner({ error }: { error: unknown }) {
-  const detail = error instanceof ApiError ? error.message : "Something went wrong. Please try again.";
-
-  return (
-    <div className="app-auth-error" role="alert">
-      <CircleAlert size={16} />
-      <span>{detail}</span>
-    </div>
-  );
-}
-
-function DashboardPage() {
-  return (
-    <PageShell>
-      <PageHeader
-        title={brand.courseLabel}
-        subtitle={`${brand.courseTerm} • ${brand.viewLabel}`}
-        actions={
-          <>
-            <button className="app-secondary-action" type="button">
-              <LayoutDashboard size={16} />
-              Open Gradebook
-            </button>
-            <button className="app-primary-action" type="button">
-              <Upload size={16} />
-              Publish All
-            </button>
-          </>
-        }
-      />
-
-      <div className="app-metric-grid">
-        <MetricCard label="Drafts" value="4" context="Items" />
-        <MetricCard label="Published" value="12" context="Live" tone="primary" />
-        <MetricCard label="Pending Review" value="8" context="Grading" tone="danger" />
-        <MetricCard label="Completed" value="45" context="Students" />
-      </div>
-
-      <div className="app-dashboard-grid">
-        <SurfacePanel className="app-table-panel">
-          <SectionHeading
-            icon={<BookOpen size={16} />}
-            title="Active Coursework"
-            actions={
-              <button className="app-inline-control" type="button">
-                Filter
-                <ChevronDown size={14} />
-              </button>
-            }
-          />
-
-          <table className="app-dense-table">
-            <thead>
-              <tr>
-                <th>Name</th>
-                <th>Type</th>
-                <th>Status</th>
-                <th>Due Date</th>
-                <th className="align-right">Submissions</th>
-              </tr>
-            </thead>
-            <tbody>
-              {assignmentRows.map((row) => (
-                <tr key={row.name}>
-                  <td className="app-table-strong">{row.name}</td>
-                  <td>{row.type}</td>
-                  <td>
-                    <StatusPill tone={row.status.tone}>{row.status.label}</StatusPill>
-                  </td>
-                  <td>{row.due}</td>
-                  <td className="align-right app-table-strong">{row.submissions}</td>
-                </tr>
-              ))}
-            </tbody>
-          </table>
-        </SurfacePanel>
-
-        <SurfacePanel className="app-feed-panel">
-          <SectionHeading icon={<Sparkles size={16} />} title="Operational Feed" />
-          <div className="app-feed-list">
-            {activityFeed.map((item) => (
-              <article key={item.title} className="app-feed-item">
-                <div className={`app-feed-icon app-feed-icon-${item.tone}`}>
-                  {item.tone === "danger" ? <CircleAlert size={16} /> : <FileCode2 size={16} />}
-                </div>
-                <div>
-                  <strong>{item.title}</strong>
-                  <p>{item.when}</p>
-                  <span>{item.detail}</span>
-                </div>
-              </article>
-            ))}
-          </div>
-        </SurfacePanel>
-      </div>
-
-      <BottomActionBar
-        leading={
-          <>
-            <span className="app-smallcaps">Quick Actions</span>
-            <button className="app-icon-button" type="button" aria-label="Search quick actions">
-              <Search size={16} />
-            </button>
-          </>
-        }
-        actions={
-          <>
-            <button className="app-inline-action" type="button">Regrade</button>
-            <button className="app-inline-action" type="button">Download CSV</button>
-          </>
-        }
-      />
-    </PageShell>
-  );
-}
-
-function AssignmentsPage() {
-  return (
-    <PageShell>
-      <PageHeader
-        eyebrow="Foundation Surface"
-        title="Assignment workflows"
-        subtitle="The design system now supports calm inspector layouts, file rails, and editorial form treatment for future instructor flows."
-        actions={
-          <Link to="/assignments/new" className="app-primary-action">
-            Open Editor
-          </Link>
-        }
-      />
-      <EmptyState
-        icon={<FolderOpen size={18} />}
-        title="Reusable assignment patterns are ready"
-        description="Issue #8 can plug actual assignment data and actions into this shared shell without restyling the workspace."
-      />
-    </PageShell>
-  );
-}
-
-function AssignmentBuilderPage() {
-  return (
-    <PageShell>
-      <PageHeader eyebrow="Assignment Editor" title="Homework 4: Linked Lists" />
-
-      <div className="app-editor-grid">
-        <div className="app-editor-main">
-          <SectionHeading
-            icon={<FileCode2 size={16} />}
-            title="Description"
-            actions={
-              <div className="app-inline-group">
-                <button className="app-inline-action app-inline-action-active" type="button">
-                  Write
-                </button>
-                <button className="app-inline-action" type="button">
-                  Preview
-                </button>
-              </div>
-            }
-          />
-
-          <SurfacePanel muted className="app-editor-block">
-            <pre>{`## Instructions
-Implement a singly linked list with the following methods:
-- append(value)
-- prepend(value)
-- delete(value)
-- find(value)
-
-### Constraints
-- Time Complexity: O(n) for searching
-- Space Complexity: O(1) for deletions
-
-Ensure all edge cases are handled (empty list, single node list).`}</pre>
-          </SurfacePanel>
-
-          <SectionHeading icon={<FolderOpen size={16} />} title="Starter Files" />
-          <div className="app-file-stack">
-            {starterFiles.map((file) => (
-              <SurfacePanel key={file.name} muted className="app-file-row">
-                <div className="app-file-main">
-                  <div className="app-file-icon">{file.icon}</div>
-                  <div>
-                    <strong>{file.name}</strong>
-                    <p>{file.meta}</p>
-                  </div>
-                </div>
-              </SurfacePanel>
-            ))}
-            <div className="app-upload-zone">
-              <Upload size={18} />
-              Upload Additional Files
-            </div>
-          </div>
-        </div>
-
-        <SurfacePanel muted className="app-editor-inspector">
-          <SectionHeading title="Metadata" />
-          <div className="app-inspector-grid">
-            <FormFieldScaffold label="Due Date">
-              <div className="app-field-value">Oct 15, 2026</div>
-            </FormFieldScaffold>
-            <FormFieldScaffold label="Points">
-              <div className="app-field-value">100</div>
-            </FormFieldScaffold>
-            <FormFieldScaffold label="Language">
-              <div className="app-field-value">Python 3.10</div>
-            </FormFieldScaffold>
-            <FormFieldScaffold label="Submission Limit" support="3 attempts">
-              <div className="app-slider-track">
-                <span />
-              </div>
-            </FormFieldScaffold>
-          </div>
-          <div className="app-inspector-meta">
-            <div className="app-meta-row">
-              <span>Visibility</span>
-              <StatusPill>Hidden</StatusPill>
-            </div>
-            <div className="app-meta-row">
-              <span>Category</span>
-              <StatusPill tone="primary">Coding</StatusPill>
-            </div>
-          </div>
-          <button className="app-verify-card" type="button">
-            <Sparkles size={22} />
-            Verify Environment
-          </button>
-        </SurfacePanel>
-      </div>
-
-      <BottomActionBar
-        leading={
-          <div className="app-status-cluster">
-            <span className="app-smallcaps">Current Status</span>
-            <strong>Draft Saving...</strong>
-          </div>
-        }
-        actions={
-          <>
-            <button className="app-inline-action" type="button">Save Draft</button>
-            <button className="app-primary-action" type="button">Publish</button>
-          </>
-        }
-      />
-    </PageShell>
-  );
-}
-
-function SubmissionsPage() {
-  return (
-    <PageShell>
-      <PageHeader
-        title="Submission queue"
-        subtitle="Shared list, status, and action-bar primitives are ready for student and staff queue flows."
-      />
-      <EmptyState
-        icon={<ListChecks size={18} />}
-        title="Submission workflows will plug in here"
-        description="The design system now supports high-density queue views and calm operational empty states."
-      />
-    </PageShell>
-  );
-}
-
-function GradebookPage() {
-  return (
-    <PageShell>
-      <div className="app-gradebook-shell">
-        <div className="app-gradebook-main">
-          <SurfacePanel muted className="app-overview-card">
-            <h3>Academic Overview</h3>
-            <p>{brand.courseLabel} • {brand.courseTerm}</p>
-            <div className="app-overview-metrics">
-              <div>
-                <span className="app-smallcaps">Average Grade</span>
-                <strong>84.2</strong>
-              </div>
-              <div>
-                <span className="app-smallcaps">Completion Rate</span>
-                <strong>92%</strong>
-              </div>
-              <div>
-                <span className="app-smallcaps">Next Deadline</span>
-                <strong>Oct 12</strong>
-              </div>
-            </div>
-          </SurfacePanel>
-
-          <div className="app-filter-row">
-            <FormFieldScaffold label="Filter by Student">
-              <div className="app-filter-pill">
-                All Students
-                <ChevronDown size={16} />
-              </div>
-            </FormFieldScaffold>
-            <FormFieldScaffold label="Assignment Status">
-              <div className="app-filter-pill">
-                All Statuses
-                <Filter size={16} />
-              </div>
-            </FormFieldScaffold>
-            <div className="app-filter-actions">
-              <button className="app-secondary-action" type="button">
-                <Download size={16} />
-                Export CSV
-              </button>
-              <button className="app-secondary-action" type="button">
-                <Settings size={16} />
-                View Settings
-              </button>
-            </div>
-          </div>
-
-          <SurfacePanel className="app-gradebook-table-panel">
-            <table className="app-gradebook-table">
-              <thead>
-                <tr>
-                  <th>Student Name</th>
-                  <th>HW 1</th>
-                  <th>HW 2</th>
-                  <th>HW 3</th>
-                  <th>HW 4</th>
-                  <th>Midterm</th>
-                  <th className="align-right">Overall</th>
-                </tr>
-              </thead>
-              <tbody>
-                {gradebookRows.map((row) => (
-                  <tr key={row.email}>
-                    <td>
-                      <div className="app-student-cell">
-                        <div className={`app-student-avatar app-student-avatar-${row.accent}`}>
-                          {row.initials}
-                        </div>
-                        <div>
-                          <strong>{row.student}</strong>
-                          <p>{row.email}</p>
-                        </div>
-                      </div>
-                    </td>
-                    {row.scores.map((score, index) => (
-                      <td key={`${row.email}-${index}`}>
-                        <strong className="app-grade-cell">{score.value}</strong>
-                        <StatusPill tone={score.tone}>{score.status}</StatusPill>
-                      </td>
-                    ))}
-                    <td className={`align-right app-overall-${row.overallTone}`}>
-                      {row.overall}
-                    </td>
-                  </tr>
-                ))}
-              </tbody>
-            </table>
-          </SurfacePanel>
-        </div>
-
-        <SurfacePanel className="app-chart-panel">
-          <SectionHeading title="Grading Distribution" />
-          <div className="app-chart-bars">
-            {gradeDistribution.map((bar) => (
-              <div key={bar.label} className="app-chart-column">
-                <div
-                  className={`app-chart-bar ${bar.accent ? "app-chart-bar-accent" : ""}`}
-                  style={{ height: bar.height }}
-                />
-                <span>{bar.label}</span>
-              </div>
-            ))}
-          </div>
-        </SurfacePanel>
-      </div>
-
-      <BottomActionBar
-        leading={
-          <div className="app-status-cluster">
-            <span className="app-smallcaps">3 students selected</span>
-            <strong>Batch actions ready</strong>
-          </div>
-        }
-        actions={
-          <>
-            <button className="app-inline-action" type="button">Message</button>
-            <button className="app-inline-action" type="button">Batch Edit</button>
-            <button className="app-primary-action" type="button">Release Selected</button>
-          </>
-        }
-      />
-    </PageShell>
-  );
-}
-
-function StudentsPage() {
-  return (
-    <PageShell>
-      <PageHeader
-        title="Roster surfaces"
-        subtitle="The shared shell, list, and metadata treatments are ready for future student-centric workflows."
-      />
-      <EmptyState
-        icon={<GraduationCap size={18} />}
-        title="Roster primitives are now part of the system"
-        description="Issue-specific student views can reuse the gradebook table rhythm, metadata labels, and sidebar shell without forking styles."
-      />
-    </PageShell>
-  );
-}
-
-function SettingsPage() {
-  return (
-    <PageShell>
-      <PageHeader
-        title="Course settings"
-        subtitle="Configuration pages can reuse the same inspector and form-field scaffolds from the assignment editor."
-      />
-      <div className="app-settings-grid">
-        <SurfacePanel muted className="app-settings-panel">
-          <SectionHeading title="Grading policy" icon={<Gauge size={16} />} />
-          <FormFieldScaffold label="Release cadence">
-            <div className="app-field-value">Manual review gate</div>
-          </FormFieldScaffold>
-          <FormFieldScaffold label="Visibility default">
-            <div className="app-field-value">Hidden until publish</div>
-          </FormFieldScaffold>
-        </SurfacePanel>
-        <SurfacePanel muted className="app-settings-panel">
-          <SectionHeading title="Operational defaults" icon={<Sparkles size={16} />} />
-          <FormFieldScaffold label="Autograde retries">
-            <div className="app-field-value">2 retries</div>
-          </FormFieldScaffold>
-          <FormFieldScaffold label="Audit retention">
-            <div className="app-field-value">Full term archive</div>
-          </FormFieldScaffold>
-        </SurfacePanel>
-      </div>
-    </PageShell>
-  );
-}
+import { queryClient, router } from "./routes/router";
 
 const root = ReactDOM.createRoot(document.getElementById("root")!);
