+++ frontend/apps/apollo/src/routes/dashboard.tsx
@@ -1,0 +1,10 @@
+import { createRoute } from "@tanstack/react-router";
+
+import { DashboardPage } from "../features/dashboard/DashboardPage";
+import { protectedLayoutRoute } from "./protected";
+
+export const dashboardRoute = createRoute({
+  getParentRoute: () => protectedLayoutRoute,
+  path: "/dashboard",
+  component: DashboardPage,
+});
