+++ frontend/apps/apollo/src/routes/root.tsx
@@ -1,0 +1,14 @@
+import type { QueryClient } from "@tanstack/react-query";
+import { Outlet, createRootRouteWithContext } from "@tanstack/react-router";
+
+export type AppContext = {
+  queryClient: QueryClient;
+};
+
+function RootRouteComponent() {
+  return <Outlet />;
+}
+
+export const rootRoute = createRootRouteWithContext<AppContext>()({
+  component: RootRouteComponent,
+});
